Split out sync and async parquet readers
