+++ arro3-io/src/parquet/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "async")]
+mod r#async;
+mod sync;
+
+#[cfg(feature = "async")]
+pub(crate) use r#async::read_parquet_async;
+pub(crate) use sync::{read_parquet, write_parquet};
