+++ arro3-io/src/parquet/sync.rs
@@ -5,7 +5,6 @@ use std::sync::Arc;
 use arrow_array::{RecordBatchIterator, RecordBatchReader};
 use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
 use parquet::arrow::arrow_writer::ArrowWriterOptions;
-use parquet::arrow::async_reader::ParquetObjectReader;
 use parquet::arrow::ArrowWriter;
 use parquet::basic::{Compression, Encoding};
 use parquet::file::properties::{WriterProperties, WriterVersion};
@@ -16,10 +15,8 @@ use pyo3::prelude::*;
 use pyo3_arrow::error::PyArrowResult;
 use pyo3_arrow::export::Arro3RecordBatchReader;
 use pyo3_arrow::input::AnyRecordBatch;
-use pyo3_arrow::{PyRecordBatchReader, PyTable};
-use pyo3_object_store::PyObjectStore;
+use pyo3_arrow::PyRecordBatchReader;
 
-use crate::error::Arro3IoResult;
 use crate::utils::{FileReader, FileWriter};
 
 #[pyfunction]
@@ -42,39 +39,6 @@ pub fn read_parquet(file: FileReader) -> PyArrowResult<Arro3RecordBatchReader> {
     // https://github.com/apache/arrow-rs/pull/5135
     let iter = Box::new(RecordBatchIterator::new(reader, arrow_schema));
     Ok(PyRecordBatchReader::new(iter).into())
-}
-
-#[pyfunction]
-#[pyo3(signature = (path, *, store))]
-pub fn read_parquet_async(
-    py: Python,
-    path: String,
-    store: PyObjectStore,
-) -> PyArrowResult<PyObject> {
-    let fut = pyo3_async_runtimes::tokio::future_into_py(py, async move {
-        Ok(read_parquet_async_inner(store.into_inner(), path).await?)
-    })?;
-
-    Ok(fut.into())
-}
-
-async fn read_parquet_async_inner(
-    store: Arc<dyn object_store::ObjectStore>,
-    path: String,
-) -> Arro3IoResult<PyTable> {
-    use futures::TryStreamExt;
-    use parquet::arrow::ParquetRecordBatchStreamBuilder;
-
-    let object_reader = ParquetObjectReader::new(store, path.into());
-    let builder = ParquetRecordBatchStreamBuilder::new(object_reader).await?;
-
-    let metadata = builder.schema().metadata().clone();
-    let reader = builder.build()?;
-
-    let arrow_schema = Arc::new(reader.schema().as_ref().clone().with_metadata(metadata));
-
-    let batches = reader.try_collect::<Vec<_>>().await?;
-    Ok(PyTable::try_new(batches, arrow_schema)?)
 }
 
 pub(crate) struct PyWriterVersion(WriterVersion);
